+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.5.1 (unreleased)
+
+* Added junkmail, an example bot that pulls random pages from an Internet
+  Archive collection.
+
 # 0.5.0
 
 Initial public release.